requirements.txt
Django==1.3.1
MySQL-python==1.2.3
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
simplejson==2.3.2
wsgiref==0.1.2
